final update

+++ src/PC/BuildMyPC/Resolution.java
@@ -1,8 +1,8 @@
 package PC.BuildMyPC;
 
 public class Resolution {
-    private double height;
-    private double width;
+    private int height;
+    private int width;
 
     public Resolution(int height, int width){
         this.height = height;
@@ -13,7 +13,7 @@ public class Resolution {
         return height;
     }
 
-    public void setHeight(double height) {
+    public void setHeight(int height) {
         this.height = height;
     }
 
@@ -21,7 +21,7 @@ public class Resolution {
         return width;
     }
 
-    public void setWidth(double width) {
+    public void setWidth(int width) {
         this.width = width;
     }
